+++ invenio_cli/helpers/cli_config.py
@@ -11,10 +11,19 @@
 """Invenio-cli configuration file."""
 
 from configparser import ConfigParser
+from functools import cached_property
 from pathlib import Path
 
 from ..errors import InvenioCLIConfigError
 from .filesystem import get_created_files
+from .package_managers import (
+    NPM,
+    PNPM,
+    UV,
+    JavascriptPackageManager,
+    Pipenv,
+    PythonPackageManager,
+)
 from .process import ProcessResponse
 
 
@@ -39,9 +48,10 @@ class CLIConfig(object):
 
         :param config_dir: Path to general cli config file.
         """
-        self.config_path = Path(project_dir) / self.CONFIG_FILENAME
+        self.project_path = Path(project_dir)
+        self.config_path = self.project_path / self.CONFIG_FILENAME
         self.config = ConfigParser()
-        self.private_config_path = Path(project_dir) / self.PRIVATE_CONFIG_FILENAME
+        self.private_config_path = self.project_path / self.PRIVATE_CONFIG_FILENAME
         self.private_config = ConfigParser()
 
         try:
@@ -49,8 +59,7 @@ class CLIConfig(object):
                 self.config.read_file(cfg_file)
         except FileNotFoundError as e:
             raise InvenioCLIConfigError(
-                "Missing '{0}' file in current directory. "
-                "Are you in the project folder?".format(e.filename),
+                f"Missing '{e.filename}' file in current directory. Are you in the project folder?",  # noqa
             )
 
         try:
@@ -60,6 +69,37 @@ class CLIConfig(object):
             CLIConfig._write_private_config(Path(project_dir))
             with open(self.private_config_path) as cfg_file:
                 self.private_config.read_file(cfg_file)
+
+    @cached_property
+    def python_package_manager(self) -> PythonPackageManager:
+        """Get python packages manager."""
+        manager_name = self.config[CLIConfig.CLI_SECTION].get("python_package_manager")
+        if manager_name == Pipenv.name:
+            return Pipenv()
+        elif manager_name == UV.name:
+            return UV()
+
+        if (self.project_path / "Pipfile").is_file():
+            return Pipenv()
+        elif (self.project_path / "pyproject.toml").is_file():
+            return UV()
+        else:
+            raise RuntimeError(
+                "Could not determine the Python package manager, please configure it."
+            )
+
+    @cached_property
+    def javascript_package_manager(self) -> JavascriptPackageManager:
+        """Get javascript packages manager."""
+        manager_name = self.config[CLIConfig.CLI_SECTION].get(
+            "javascript_package_manager"
+        )
+        if manager_name == NPM.name:
+            return NPM()
+        elif manager_name == PNPM.name:
+            return PNPM()
+
+        return NPM()
 
     def get_project_dir(self):
         """Returns path to project directory."""
@@ -112,14 +152,22 @@ class CLIConfig(object):
 
     def get_search_port(self):
         """Returns the search port."""
-        return self.config[CLIConfig.COOKIECUTTER_SECTION].get("search_port", "9200")
+        return self.private_config[CLIConfig.CLI_SECTION].get("search_port", "9200")
 
     def get_search_host(self):
         """Returns the search host."""
-        return self.config[CLIConfig.COOKIECUTTER_SECTION].get(
+        return self.private_config[CLIConfig.CLI_SECTION].get(
             "search_host",
             "localhost",
         )
+
+    def get_web_port(self):
+        """Returns web port."""
+        return self.private_config[CLIConfig.CLI_SECTION].get("web_port", "5000")
+
+    def get_web_host(self):
+        """Returns web host."""
+        return self.private_config[CLIConfig.CLI_SECTION].get("web_host", "127.0.0.1")
 
     def get_db_type(self):
         """Returns the database type (mysql, postgresql)."""